payments_api/requirements.txt
pika
python-dotenv
flask
flask-cors
requests
rsa